requirements.txt
fastapi[standard]
python-dotenv
pydantic-settings
alembic
sqlmodel
passlib[argon2]
python-jose[cryptography]
fastapi-mail

